feat: refactor SettingsPage into modular tab components
Extract each settings section into a dedicated component:
GeneralTab, SecurityTab, UiTab, TelegramTab, TunnelsTab,
TerminalTab, LogAlertsTab, AiTab, and shared.tsx utilities.
SettingsPage becomes a thin shell that renders the active tab.
